+++ sql/update_2.8.0-2.9.0.sql
@@ -1,0 +1,1 @@
+ALTER TABLE llx_scrumproject_scrumuserstory ADD column complete_task_on_us_done tinyint NOT NULL DEFAULT 0;